Auto scrolling user comment list & more
Updated the interface to make it user friendly, making it simpler by replacing the user comment section about the changes to a flex box, that scrolls automatically and can be paused (to control it manually or just to read), and can be reset.

Made a other few changes. Check PR diff.